[CLN] account: split and extact bill_upload logic
This commit split bill_upload files to specific files and
extract common part of upload logic to be able to reuse
it in other models like in sale.order including views.

task-3791197
