+++ addons/account/__manifest__.py
@@ -105,7 +105,7 @@ You could use this simplified accounting in case you work with an (external) acc
             'account/static/src/scss/account_multi_ledger.scss',
             'account/static/src/components/**/*',
             'account/static/src/services/*.js',
-            'account/static/src/views/*.js',
+            'account/static/src/views/**/*',
             'account/static/src/js/tours/account.js',
             'account/static/src/js/search/search_bar/search_bar.js',
             'account/static/src/helpers/*.js',
