+++ addons/account/static/src/components/account_file_uploader/account_file_uploader.scss
@@ -1,0 +1,7 @@
+.o_widget_account_file_uploader {
+  .oe_kanban_action_button {
+    a {
+      color: var(--btn-color);
+    }
+  }
+}
