--- addons/account/static/src/components/bills_upload/bills_upload.scss
@@ -1,66 +1,0 @@
-.o_drop_area {
-    width: calc(100% + 2px);
-    height: calc(100% + 2px);
-    position: absolute;
-    background-color: #AAAA;
-    z-index: 3;
-    left: -1px;
-    top: -1px;
-    i {
-      justify-content: center;
-      display: flex;
-      align-items: center;
-      height: 100%;
-    }
-    .upload_badge {
-      top: 47% !important;
-      left: 53% !important
-    }
-}
-.o_account_dashboard_kanban_view {
-  .o_drop_area {
-    background-color: $o-view-background-color;
-    border: 1px dashed $border-color;
-    &.drag_to_card {
-      border-color: $o-brand-primary;
-      background-color: #F2EDF0;
-    }
-  }
-}
-
-.o_view_nocontent {
-  .o_nocontent_help:has(> .bill_guide_container) {
-    min-width: 65vw;
-  }
-}
-.bill_guide_container {
-  @include media-breakpoint-up(sm) {
-    min-width: 400px;
-  }
-
-  .bill_guide_left, .bill_guide_right {
-    width: 45%;
-  }
-
-  .separator_wrapper {
-    width: 10%;
-  }
-}
-
-.file_upload_kanban_action_a {
-  @include o-kanban-dashboard-dropdown-link($link-padding-gap: $o-kanban-dashboard-dropdown-complex-gap);
-}
-
-.o_widget_account_file_uploader {
-  .oe_kanban_action_button {
-    a {
-      color: var(--btn-color);
-    }
-  }
-}
-
-.account_drag_drop_btn {
-  border-style: dashed !important;
-  border-color: $o-brand-primary;
-  background-color: #F2EDF0;
-}
