+++ addons/account/static/src/components/document_file_uploader/document_file_uploader.xml
@@ -1,0 +1,29 @@
+<templates>
+
+    <t t-name="account.DocumentFileUploader">
+        <div t-att-class="divClass">
+            <FileUploader
+                acceptedFileExtensions="props.acceptedFileExtensions"
+                fileUploadClass="'document_file_uploader'"
+                multiUpload="true"
+                onUploaded.bind="onFileUploaded"
+                onUploadComplete.bind="onUploadComplete">
+                <t t-set-slot="toggler">
+                    <t t-slot="toggler"/>
+                </t>
+                <t t-slot="default"/>
+            </FileUploader>
+        </div>
+    </t>
+
+    <t t-name="account.DocumentViewUploadButton">
+        <DocumentFileUploader resModel="props.resModel">
+            <t t-set-slot="toggler">
+                <button type="button" class="btn btn-secondary" data-hotkey="b">
+                    Upload
+                </button>
+            </t>
+        </DocumentFileUploader>
+    </t>
+
+</templates>
