+++ addons/account/static/src/components/upload_drop_zone/upload_drop_zone.scss
@@ -1,0 +1,20 @@
+.o_drop_area {
+    width: calc(100% + 2px);
+    height: calc(100% + 2px);
+    position: absolute;
+    background-color: grey;
+    z-index: 3;
+    left: -1px;
+    top: -1px;
+    i {
+      justify-content: center;
+      display: flex;
+      align-items: center;
+      height: 100%;
+    }
+    .upload_badge {
+      top: 50% !important;
+      left: 50% !important;
+      transform: translate(-50%,-50%);
+    }
+}
