+++ addons/account/static/src/views/account_dashboard_kanban/account_dashboard_kanban_record.xml
@@ -1,0 +1,27 @@
+<templates>
+
+    <t t-name="account.DashboardKanbanRecord">
+        <div
+            role="article"
+            t-att-class="getRecordClasses()"
+            t-att-data-id="props.canResequence and props.record.id"
+            t-att-tabindex="props.record.model.useSampleModel ? -1 : 0"
+            t-on-click="onGlobalClick"
+            t-on-dragenter.stop.prevent="() => dropzoneState.visible = true"
+            t-ref="root">
+            <AccountFileUploader t-if="allowDrop" record="props.record">
+                <t t-set-slot="default">
+                    <UploadDropZone t-props="dropzoneProps"/>
+                    <t t-call="{{ templates[this.constructor.KANBAN_BOX_ATTRIBUTE] }}"
+                        t-call-context="renderingContext"/>
+                </t>
+            </AccountFileUploader>
+            <div t-else="" t-att-class="'oe_kanban_color_' + props.record.data.color">
+                <t t-call="{{ templates[this.constructor.KANBAN_BOX_ATTRIBUTE] }}"
+                    t-call-context="renderingContext"/>
+            </div>
+            <t t-call="{{ this.constructor.menuTemplate }}"/>
+        </div>
+    </t>
+
+</templates>
