+++ addons/account/static/src/views/account_dashboard_kanban/account_dashboard_kanban_renderer.xml
@@ -1,0 +1,12 @@
+<templates>
+
+    <t t-name="account.DashboardKanbanRenderer" t-inherit="web.KanbanRenderer" t-inherit-mode="primary">
+        <xpath expr="//div[@t-ref='root']" position="attributes">
+            <attribute name="t-on-dragenter.prevent">kanbanDragEnter</attribute>
+            <attribute name="t-on-dragleave">kanbanDragLeave</attribute>
+            <attribute name="t-on-drop.prevent">kanbanDragDrop</attribute>
+            <attribute name="t-on-dragover.prevent">() =&gt; {}</attribute>
+        </xpath>
+    </t>
+
+</templates>
