+++ addons/account/static/src/views/account_move_kanban/account_move_kanban_view.js
@@ -1,0 +1,11 @@
+import { registry } from "@web/core/registry";
+import { fileUploadKanbanView } from "../file_upload_kanban/file_upload_kanban_view";
+import { AccountMoveKanbanController } from "./account_move_kanban_controller";
+
+export const accountMoveUploadKanbanView = {
+    ...fileUploadKanbanView,
+    Controller: AccountMoveKanbanController,
+    buttonTemplate: "account.AccountMoveKanbanView.Buttons",
+};
+
+registry.category("views").add("account_documents_kanban", accountMoveUploadKanbanView);
