+++ addons/account/static/src/views/account_move_list/account_move_list_controller.xml
@@ -1,0 +1,9 @@
+<templates>
+
+    <t t-name="account.AccountMoveListView.Buttons" t-inherit="account.FileuploadListView.Buttons" t-inherit-mode="primary">
+        <xpath expr="//t[@t-call='account.DocumentViewUploadButton']" position="replace">
+            <t t-if="showUploadButton" t-call="account.AccountViewUploadButton"/>
+        </xpath>
+    </t>
+
+</templates>
