+++ addons/account/static/src/views/file_upload_kanban/file_upload_kanban_controller.js
@@ -1,0 +1,9 @@
+import { KanbanController } from "@web/views/kanban/kanban_controller";
+import { DocumentFileUploader } from "@account/components/document_file_uploader/document_file_uploader";
+
+export class FileUploadKanbanController extends KanbanController {
+    static components = {
+        ...KanbanController.components,
+        DocumentFileUploader,
+    };
+}
