+++ addons/account/static/src/views/file_upload_kanban/file_upload_kanban_controller.xml
@@ -1,0 +1,9 @@
+<templates>
+
+    <t t-name="account.FileuploadKanbanView.Buttons" t-inherit="web.KanbanView.Buttons" t-inherit-mode="primary">
+        <xpath expr="//div[hasclass('o_cp_buttons')]" position="inside">
+            <t t-call="account.DocumentViewUploadButton"/>
+        </xpath>
+    </t>
+
+</templates>
