+++ addons/account/static/src/views/file_upload_kanban/file_upload_kanban_renderer.js
@@ -1,0 +1,16 @@
+import { KanbanRenderer } from "@web/views/kanban/kanban_renderer";
+import { UploadDropZone } from "@account/components/upload_drop_zone/upload_drop_zone";
+import { useState } from "@odoo/owl";
+
+export class FileUploadKanbanRenderer extends KanbanRenderer {
+    static template = "account.FileUploadKanbanRenderer";
+    static components = {
+        ...KanbanRenderer.components,
+        UploadDropZone,
+    };
+
+    setup() {
+        super.setup();
+        this.dropzoneState = useState({ visible: false });
+    }
+}
