+++ addons/account/static/src/views/file_upload_kanban/file_upload_kanban_view.js
@@ -1,0 +1,13 @@
+import { registry } from "@web/core/registry";
+import { kanbanView } from "@web/views/kanban/kanban_view";
+import { FileUploadKanbanController } from "./file_upload_kanban_controller";
+import { FileUploadKanbanRenderer } from "./file_upload_kanban_renderer";
+
+export const fileUploadKanbanView = {
+    ...kanbanView,
+    Controller: FileUploadKanbanController,
+    Renderer: FileUploadKanbanRenderer,
+    buttonTemplate: "account.FileuploadKanbanView.Buttons",
+};
+
+registry.category("views").add("file_upload_kanban", fileUploadKanbanView);
