+++ addons/account/static/src/views/file_upload_list/file_upload_list_controller.js
@@ -1,0 +1,9 @@
+import { ListController } from "@web/views/list/list_controller";
+import { DocumentFileUploader } from "@account/components/document_file_uploader/document_file_uploader";
+
+export class FileUploadListController extends ListController {
+    static components = {
+        ...ListController.components,
+        DocumentFileUploader,
+    };
+};
