+++ addons/account/static/src/views/file_upload_list/file_upload_list_controller.xml
@@ -1,0 +1,9 @@
+<templates>
+
+    <t t-name="account.FileuploadListView.Buttons" t-inherit="web.ListView.Buttons" t-inherit-mode="primary">
+        <xpath expr="//div[hasclass('o_list_buttons')]" position="inside">
+            <t t-call="account.DocumentViewUploadButton"/>
+        </xpath>
+    </t>
+
+</templates>
